+++ database/seeders/DatabaseSeeder.php
@@ -25,5 +25,7 @@ class DatabaseSeeder extends Seeder
             'email' => 'test@example.com',
             'password' => bcrypt('password'),
         ]);
+
+        $this->call(DemoBusinessSeeder::class);
     }
 }